MoeChat-main/requirements.txt
torch
torchaudio
sounddevice
soundfile
numpy<2.0
scipy
jionlp
ruamel.yaml
modelscope
faiss-cpu
funasr
scipy
addict
datasets<3
simplejson
transformers
shortuuid
filetype
b64
Pillow
sortedcontainers
fastapi
uvicorn
rich
pysilero
IPython
praat-parselmouth
